refactor: move functions to its own file
