+++ crm/api/event.py
@@ -14,7 +14,6 @@ to events that don't have custom notifications set. This ensures all events can 
 reminders even if users don't configure them individually.
 """
 
-import re
 from datetime import datetime, timedelta
 
 import frappe
@@ -371,187 +370,3 @@ def _format_time_remaining(before_value, interval):
 def _send_system_notification(notification):
 	"""Send system notification for an event"""
 	frappe.publish_realtime("event_notification", notification)
-
-
-TASK_REMINDER_OFFSETS = {
-	"hours": [{"value": 1, "label": "1 hour"}],
-	"days": [{"value": 1, "label": "1 day"}],
-}
-
-
-def send_hourly_task_reminders():
-	_process_task_reminders("hours")
-
-
-def send_daily_task_reminders():
-	_process_task_reminders("days")
-
-
-def _process_task_reminders(interval: str):
-	"""
-	Process task reminders for a given interval.
-
-	Queries all open CRM Tasks with a due_date and assigned_to user,
-	calculates trigger windows, deduplicates, and sends both in-app
-	and email notifications.
-
-	Args:
-		interval (str): 'hours' or 'days'
-	"""
-
-	current_time = now_datetime()
-	offsets = TASK_REMINDER_OFFSETS.get(interval, [])
-	if not offsets:
-		return
-
-	tasks = frappe.get_all(
-		"CRM Task",
-		filters={
-			"status": ["not in", ["Done", "Canceled"]],
-			"due_date": ["is", "set"],
-			"assigned_to": ["is", "set"],
-		},
-		fields=[
-			"name",
-			"title",
-			"due_date",
-			"assigned_to",
-			"priority",
-			"status",
-			"reference_doctype",
-			"reference_docname",
-		],
-	)
-
-	for task in tasks:
-		for offset in offsets:
-			try:
-				_process_single_task_reminder(task, offset, interval, current_time)
-			except Exception as e:
-				frappe.log_error(f"Error processing {interval} reminder for task {task.name}: {e!s}")
-
-
-def _process_single_task_reminder(task: dict, offset: dict, interval: str, current_time: datetime):
-	due_date = task.due_date
-	if not isinstance(due_date, datetime):
-		due_date = frappe.utils.get_datetime(due_date)
-
-	if interval == "days":
-		target_date = add_to_date(due_date, days=-offset["value"]).date()
-		if current_time.date() != target_date:
-			return
-	else:
-		interval_kwargs = _get_interval_kwargs(interval, offset["value"])
-		trigger_time = add_to_date(due_date, **interval_kwargs)
-
-		window = _get_trigger_window_duration(interval)
-		window_start = add_to_date(trigger_time, **{k: -v for k, v in window.items()})
-		window_end = add_to_date(trigger_time, **window)
-		if not (window_start <= current_time <= window_end):
-			return
-
-	reminder_key = f"task_reminder_{interval}_{offset['value']}"
-	existing = frappe.db.exists(
-		"CRM Notification",
-		{
-			"type": "Task",
-			"notification_type_doctype": "CRM Task",
-			"notification_type_doc": str(task.name),
-			"to_user": task.assigned_to,
-			"message": ["like", f"%{reminder_key}%"],
-		},
-	)
-	if existing:
-		return
-
-	# Send in-app notification
-	_send_task_inapp_notification(task, offset, reminder_key)
-
-	# Send email notification
-	_send_task_email_notification(task, offset)
-
-
-def _send_task_inapp_notification(task: dict, offset: dict, reminder_key: str):
-	notification_text = f"""
-		<div class="mb-2 leading-5 text-ink-gray-5">
-			Task <span class="font-medium text-ink-gray-9">{task.title}</span>
-			is due in {offset["label"]}.</span>
-		</div>
-	"""
-
-	reference_doctype = (
-		task.reference_doctype if task.reference_doctype and task.reference_docname else "CRM Task"
-	)
-	reference_name = (
-		task.reference_docname if task.reference_doctype and task.reference_docname else str(task.name)
-	)
-
-	values = frappe._dict(
-		doctype="CRM Notification",
-		from_user=task.assigned_to,
-		to_user=task.assigned_to,
-		type="Task",
-		message=f"Task '{task.title}' is due in {offset['label']}. [{reminder_key}]",
-		notification_text=notification_text,
-		notification_type_doctype="CRM Task",
-		notification_type_doc=str(task.name),
-		reference_doctype=reference_doctype,
-		reference_name=reference_name,
-	)
-
-	frappe.get_doc(values).insert(ignore_permissions=True)
-
-
-def _strip_root_p_tags(text: str):
-	if not text:
-		return text
-	text = str(text).strip()
-	if text.startswith("<p>") and text.endswith("</p>"):
-		text = text[3:-4]
-		text = re.sub(r"</p>\s*<p>", "<br><br>", text)
-	return text
-
-
-def _format_due_date_for_email(due_date: str):
-	dt = frappe.utils.get_datetime(due_date)
-	am_pm = "AM" if dt.hour < 12 else "PM"
-	hour = dt.hour % 12 or 12
-	return f"{dt.strftime('%b')} {dt.day}, {dt.year}, {hour}:{dt.strftime('%M')} {am_pm}"
-
-
-def _send_task_email_notification(task: dict, offset: dict):
-	recipient = task.assigned_to
-	if not recipient:
-		return
-
-	subject = f"Task Reminder: {task.title} — due in {offset['label']}"
-
-	assigned_to_name = frappe.db.get_value("User", recipient, "full_name") or recipient
-	task_doc = frappe.get_doc("CRM Task", task.name)
-	brand_logo = frappe.db.get_single_value("FCRM Settings", "brand_logo") or "/assets/crm/images/logo.png"
-
-	site_url = frappe.utils.get_url()
-	task_url = f"{site_url}/crm/tasks/view/list?open={task.name}"
-
-	# nosemgrep
-	message = frappe.render_template(
-		"crm/templates/emails/task_reminder.html",
-		{
-			"task_title": task.title,
-			"due_date": _format_due_date_for_email(task.due_date),
-			"time_remaining": offset["label"],
-			"assigned_to_name": assigned_to_name,
-			"description": _strip_root_p_tags(task_doc.description),
-			"task_url": task_url,
-			"brand_logo": brand_logo,
-		},
-	)
-
-	frappe.sendmail(
-		recipients=[recipient],
-		subject=subject,
-		message=message,
-		reference_doctype="CRM Task",
-		reference_name=str(task.name),
-		now=True,
-	)
